Add test to ensure help examples are valid expressions
